+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ### Changes
 
 * [#6492](https://github.com/rubocop-hq/rubocop/issues/6492): Auto-correct chunks of comment lines in `Layout/CommentIndentation` to avoid unnecessary iterations for `rubocop -a`. ([@jonas054][])
+* [#6519](https://github.com/rubocop-hq/rubocop/issues/6519): Allow `Style/BlockComments` with apiDoc documentation. ([@anthony-robin][])
 
 ## 0.60.0 (2018-10-26)